Block created by full node

+++ src/types/block_header.py
@@ -1,13 +1,13 @@
 from blspy import PrependSignature
 from src.util.streamable import streamable
-from src.util.ints import uint32
+from src.util.ints import uint64
 from src.types.sized_bytes import bytes32
 
 
 @streamable
 class BlockHeaderData:
     previous_header_hash: bytes32
-    timestamp: uint32
+    timestamp: uint64
     filter_hash: bytes32
     proof_of_space_hash: bytes32
     body_hash: bytes32
